Removed unecessary UI scripts and combined indoor and outdoor UIs, instead they will alternate depending on scene
