+++ IMSE_UTARide/Assets/Art/Textures/UI_Textures/frame_light.afphoto.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 2a15fef1e83553d45a48641de4bd625e
-PrefabImporter:
+guid: 7156eb20e7e59e94cbada7e7510d9553
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
