--- IMSE_UTARide/Assets/Prefabs/UI_Prefab/UI_GameObjects_Indoor.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 50e384d2ee821ca41b5599db838ec4ab
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
