--- IMSE_UTARide/Assets/Scripts/UI_Scripts/SpawnCube.cs
@@ -1,13 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class SpawnCube : MonoBehaviour
-{
-    public GameObject cubePrefab;
-    public GameObject cubeSpawner;
-    public void Cube()
-    {
-        Instantiate(cubePrefab, cubeSpawner.transform.position,Quaternion.identity);
-    }
-}
