--- IMSE_UTARide/Assets/Scripts/UI_Scripts/SpawnCube.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e09a901d274e3064cb6c7b7988c12589
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
