--- IMSE_UTARide/Assets/Scripts/UI_Scripts/UI_MenuController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5aba04728e099fa40a26f06d6e94b6f4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
