--- IMSE_UTARide/Assets/Scripts/UI_Scripts/UI_UserInput.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 927604217007a2a4495454bd5d3ed57f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
